refactor(fixture): move `ElementHandle` based fixture into its own module
